Time: 21 ms (85.99%), Space: 16.2 MB (97.1%) - LeetHub

+++ 0155-min-stack/0155-min-stack.cpp
@@ -30,14 +30,14 @@ public:
     }
     
     int top() {
-        if (stk.top() >= mini) return stk.top();
+        if (stk.top() >= mini) return (int)stk.top();
         else{
-            return mini;
+            return (int)mini;
         }
     }
     
     int getMin() {
-        return mini;
+        return (int)mini;
     }
 };
